Trick for python vs. python3 isolate language
If a component is written in Python, its language value is "python".
If an isolate starts Python (3 by default), its value is "python3".

As they are different, the distributor requests the creation of a new isolate of language "python".
This commit ignores language differences as a workaround: this could lead to errors later (be careful).